pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prcb-checks"
version = "0.1.2"
description = "Helper command to invoke GitHub Checks API when AWS CodeBuild is executed from GitHub via AWS CodeConnections"
authors = [
    {name = "neruneruo", email = "rkubota.neru@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3",
    "cryptography",
    "PyJWT",
    "requests",
]

[project.urls]
repository = "https://github.com/neruneruo/prcb-checks"

[project.scripts]
prcb-checks = "prcb_checks.main:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock"
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["images"]
